website: Add max-width to the main container
Adding max-width to the main container of the website makes it display
better on large (wide) screens. Before this change, the website would
stretch to the entire width of the screen. Since it was not designed for
wide screens, this causes some layout and readability issues (e.g., nav
bar /menu moves down and "leaves" its container and a lot of empty space
is created between elements.

Reviewed by:		0mp
Differential Revision:	https://reviews.freebsd.org/D45801