website: Replace RSS icon asset with higher resolution ones
The RSS icon on the website is visibly low resolution. This change adds
a high resolution version of the icon and replaces it across the
website.

The new icon is double the size. Because there are instances where the
icon's "physical" size cannot be changed (because it's used in a
`list-style-image` CSS property) an additional version of the icon is
added so that the layout of those is not affected, but the asset's
quality is still improved.

The icon is renamed from ico_rss.png to rss_icon.png since replacing an
image file without renaming it was causing preview issues, and since the
new name is a slightly better one.

Reviewed by:		0mp
Differential Revision:	https://reviews.freebsd.org/D45806
Differential Revision:	https://reviews.freebsd.org/D45810