website: Visually rearrange main buttons
The current placement of the "New To FreeBSD?" button is slightly
chaotic, as it's not visually aligned with anything else, and uses a
background color not used anywhere else. The "Download FreeBSD" button
is also visually misaligned and uses a font size not used anywhere else.

This proposed change puts the two buttons together, aligns their left
edge with the "Supported releases" section, adjusts width to make sure
the titles fit on the same page, and aligns vertically with
`beastie.png`. It unifies their font size at 1.3em (which is also used
elsewhere) and changes the background color of "New To (..)" button to
be lighter and the same as used below.

Reviewed by:		0mp
Differential Revision:	https://reviews.freebsd.org/D45798